LTS minimization algorithm implemented
The minimization is performed using the strong bisimulation
equivalence concept (e.g. tau actions are treated like the regular
actions).

For now the user cannot really use it, since the minimization is
performed on a local Lts that doesn't go in the 'processes' table.
Moreover, verbose output is output on the shell.

These user-interface fixes will be addressed in future patches.