add 'clone' argument to process_ref_translate()
Before this commit the process_ref_translate() function always
cloned the LTS stored in the 'processes' table.

This is actually necessary (in the current code) only when the LTS
is used as a building block for another LTS (e.g.
ProcessRefNode::translate()).

In all the other cases the clone is not necessary or even not
desiderable. In the shell minimize command, as en example, we
want to minimize the LTS stored in the 'processes' table, and not
a clone of it.