minimization bug fixed
The helper functions children_to_*_set have been moved inside the
Lts class. By explicitely clearing the output set (s.clear()), a
bug has been found: The helper function assumed the set to be empty,
but this was not guaranteed by the caller. The minimized output for
the test case 1 was therefore incorrect and has been replaced.